### 2020-01-07 #1996 @whedon
[PRE REVIEW]: vfi: Classification by Voting Feature Intervals 
**Submitting author:** @chkoar (<a href="http://orcid.org/0000-0002-5021-1442">Christos K. Aridas</a>)
**Repository:** <a href="https://github.com/chkoar/vfi" target ="_blank">https://github.com/chkoar/vfi</a>
**Version:** 0.0.1
