### 2020-01-07 9f0707a
Validate parameters for friendlier error messages
