### 2020-01-07 60c6264
Update publish task
