### 2020-01-08 #2005 @whedon
[REVIEW]: vfi: Classification by Voting Feature Intervals 
**Submitting author:** @chkoar (<a href="http://orcid.org/0000-0002-5021-1442">Christos K. Aridas</a>)
**Repository:** <a href="https://github.com/chkoar/vfi" target ="_blank">https://github.com/chkoar/vfi</a>
**Version:** 0.0.1
