### 2020-01-09 4d138f1
Update citations
