+++ paper/paper.bib
@@ -10,7 +10,8 @@ publisher="Springer Berlin Heidelberg",
 address="Berlin, Heidelberg",
 pages="85--92",
 abstract="A new classification algorithm called VFI (for Voting Feature Intervals) is proposed. A concept is represented by a set of feature intervals on each feature dimension separately. Each feature participates in the classification by distributing real-valued votes among classes. The class receiving the highest vote is declared to be the predicted class. VFI is compared with the Naive Bayesian Classifier, which also considers each feature separately. Experiments on real-world datasets show that VFI achieves comparably and even better than NBC in terms of classification accuracy. Moreover, VFI is faster than NBC on all datasets.",
-isbn="978-3-540-68708-5"
+isbn="978-3-540-68708-5",
+doi = {10.1007/3-540-62858-4_74}
 }
 
 @article{pedregosa2011,
@@ -26,11 +27,19 @@ isbn="978-3-540-68708-5"
 }
 
 @article{hall2009,
-  author = {Mark Hall and Eibe Frank and Geoffrey Holmes and Bernhard Pfahringer and Peter Reutemann and Ian H. Witten},
-  title = {The {WEKA} data mining software: an update},
-  journal = {SIGKDD Explorations},
-  volume = 11,
-  number = 1,
-  year = 2009,
-  pages = {10-18}
-} 
+ author = {Hall, Mark and Frank, Eibe and Holmes, Geoffrey and Pfahringer, Bernhard and Reutemann, Peter and Witten, Ian H.},
+ title = {The WEKA Data Mining Software: An Update},
+ year = {2009},
+ issue_date = {November 2009},
+ publisher = {Association for Computing Machinery},
+ address = {New York, NY, USA},
+ volume = {11},
+ number = {1},
+ issn = {1931-0145},
+ url = {https://doi.org/10.1145/1656274.1656278},
+ doi = {10.1145/1656274.1656278},
+ journal = {SIGKDD Explor. Newsl.},
+ month = nov,
+ pages = {10–18},
+ numpages = {9}
+}
