### 2020-01-09 073fb38
Add stronger validation for n_bins parameter
