### 2020-01-10 8dfafc1
Fix the example with the decision boundaries
