### 2020-01-10 #2005 @arfon
Hi all, my sincere apologies for not catching this sooner but I'm afraid we're going to have to reject this submission. While it looks to be a useful little package, this falls into our ['minor utility'](https://joss.readthedocs.io/en/latest/submitting.html) category.

@chkoar @mbobra - apologies for not catching this at the `pre-review` stage of the process when editorial decisions in JOSS like this are meant to be made.

@jakryd @dflemin3 - thank you for agreeing to review this submission. From the looks of things it appears that you've not started your reviews yet so hopefully we've not wasted too much of your time.
