### 2020-01-10 #2005 @mbobra
@arfon No worries, I understand. @danielskatz asked me to edit this submission and I was happy to do so.

I do want to point out that @jakryd @dflemin3 were both excited about reviewing for JOSS and @dflemin3 even went through the trouble of reading the JOSS Editorial Guide to prep for this review. So I hope we can be a little better about identifying minor submissions before we solicit reviewers. I do understand that mix-ups happen sometimes.

A big thank you to @jakryd @dflemin3 for volunteering so readily on top of your already busy workload ☀️
