### 2020-01-11 #2005 @chkoar
> While it looks to be a useful little package, this falls into our 'minor utility' category.

Dear @arfon, @mbobra  and @danielskatz,

I respectively disagree that the package is a "minor utility" and I would incline you to reconsider. Currently, it contains a real world classifier and even at this stage it can be used in research works. My main motivation, in order to identify if the package at this stage could be considered as minor utility and make the submission, were utilities like [this](https://joss.theoj.org/papers/10.21105/joss.01593) and [this](https://joss.theoj.org/papers/10.21105/joss.00124) that have been accepted for publication in JOSS. At the time of publication, the former had one public function while the latter had two public functions.

Best,
Chris
