### 2020-01-12 #2005 @arfon
Dear @chkoar, firstly I want to apologize again for how this submission was handled. This submission was flagged for rejection early on by a handling editor but got lost in a transition between editors over the holiday season.

To address your concern about this submission being treated differently to those of a similar size/scope: the JOSS editorial team has recently carried out a significant review of our scope as a journal (i.e. what submissions are considered in/out of scope) and your paper and software was an example of us applying these criteria. Should these other two submissions be submitted today we would reject them too.
