### 2020-01-13 #2005 @chkoar
Dear @arfon,

Hopefully, this will be my last reply.

> the JOSS editorial team has recently carried out a significant review of our scope as a journal

As a feedback, you should state how they differ with the past? As I said previously my motivation for submission were past accepted papers and JOSS' [about](https://joss.theoj.org/about) page.

> To address your concern about this submission being treated differently to those of a similar size/scope

Please, don't get me wrong. The whole thing is not about my rejection. But, it seems strange (apart from your change of scope) that an open journal like this has inconsistent decisions about submissions.

For example:
You accept [this](https://joss.theoj.org/papers/10.21105/joss.00124) (which was a clear example of minor utility in my honest opinion), later, as handling editor, you promote [this](https://github.com/openjournals/joss-reviews/issues/804) submission from pre-review stage to [review](https://github.com/openjournals/joss-reviews/issues/928) and you reject it as minor utility. Then again you accept [this](https://joss.theoj.org/papers/10.21105/joss.01593) submission but you reject [this](https://github.com/openjournals/joss-reviews/issues/1966) one for the same reasons you accepted the previous ones.

In [this](https://joss.theoj.org/about) page of the JOSS site you state

```
JOSS publishes articles about research software. This definition includes software that: solves complex modeling problems in a scientific context (physics, mathematics, biology, medicine, social science, neuroscience, engineering); supports the functioning of research instruments or the execution of research experiments; extracts knowledge from large data sets; offers a mathematical library; or similar.

JOSS submissions must:
    Be open source (i.e., have an OSD-compliant license).
    Have an obvious research application.
    Be feature-complete (no half-baked solutions) and be designed for maintainable extension (not one-off modifications).
    Not be minor 'utility' packages, such as 'thin' API clients.
```
In contrast to the other rejected submissions that I mentioned, I believe that my submission was fulfilling the definition of research software and all the above submission musts.

In my honest opinion, since a submission has passed by the hands of an Associate Editor-in-Chief and an Editor and the reviewers were found, the submission should go for a review round. In my eyes their role was nullified. At the end of the day the EiC decides.

Best,
Chris
